% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/aggregateToComposition.R
\name{aggregateToComposition}
\alias{aggregateToComposition}
\alias{aggregateToComposition,SingleCellExperiment-method}
\title{Aggregate a SingleCellExperiment to a patient x cell-type composition}
\usage{
aggregateToComposition(sce, patient_col, celltype_col, ...)

\S4method{aggregateToComposition}{SingleCellExperiment}(sce, patient_col, celltype_col, ...)
}
\arguments{
\item{sce}{A \link[SingleCellExperiment:SingleCellExperiment]{SingleCellExperiment::SingleCellExperiment}.}

\item{patient_col}{Column in \code{colData(sce)} identifying patients.}

\item{celltype_col}{Column in \code{colData(sce)} identifying cell types.}

\item{...}{Unused.}
}
\value{
A \link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment::SummarizedExperiment} with composition
proportions in \code{assay()} and patient metadata in \code{colData}.
}
\description{
Aggregate a SingleCellExperiment to a patient x cell-type composition
}
\section{Methods (by class)}{
\itemize{
\item \code{aggregateToComposition(SingleCellExperiment)}: Method for
\link[SingleCellExperiment:SingleCellExperiment]{SingleCellExperiment::SingleCellExperiment}.

}}
\examples{
# Scaffold: method body lands in Phase 3.2b.
showMethods("aggregateToComposition")
}
