% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/accessors.R
\name{hFractions}
\alias{hFractions}
\alias{hFractions,TMESplitResult-method}
\title{Accessor: patient-level program activities (soft assignment)}
\usage{
hFractions(x)

\S4method{hFractions}{TMESplitResult}(x)
}
\arguments{
\item{x}{A \link{TMESplitResult}.}
}
\value{
A numeric matrix of shape \verb{n_patients x k_total}; rows sum to 1.
}
\description{
Accessor: patient-level program activities (soft assignment)
}
\section{Methods (by class)}{
\itemize{
\item \code{hFractions(TMESplitResult)}: Method for \link{TMESplitResult}.

}}
\examples{
res <- new("TMESplitResult")
hFractions(res)
}
