% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotting.R
\name{plotPrograms}
\alias{plotPrograms}
\alias{plotPrograms,TMESplitResult-method}
\title{Plot program loadings (W) as a heatmap}
\usage{
plotPrograms(x, ...)

\S4method{plotPrograms}{TMESplitResult}(x, ...)
}
\arguments{
\item{x}{A \link{TMESplitResult}.}

\item{...}{Passed to \link[ComplexHeatmap:Heatmap]{ComplexHeatmap::Heatmap}.}
}
\value{
A \link[ComplexHeatmap:Heatmap]{ComplexHeatmap::Heatmap} object (invisibly), drawn as a side
effect.
}
\description{
Plot program loadings (W) as a heatmap
}
\section{Methods (by class)}{
\itemize{
\item \code{plotPrograms(TMESplitResult)}: Method for \link{TMESplitResult}.

}}
\examples{
# Scaffold: body lands in Phase 3.2c.
res <- new("TMESplitResult")
tryCatch(plotPrograms(res), error = function(e) message(conditionMessage(e)))
}
