% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotting.R
\name{plotSignificance}
\alias{plotSignificance}
\alias{plotSignificance,TMESplitResult-method}
\title{Plot significance (permutation null vs observed)}
\usage{
plotSignificance(x, ...)

\S4method{plotSignificance}{TMESplitResult}(x, ...)
}
\arguments{
\item{x}{A \link{TMESplitResult}.}

\item{...}{Extra aesthetics.}
}
\value{
A \link[ggplot2:ggplot]{ggplot2::ggplot} object.
}
\description{
Plot significance (permutation null vs observed)
}
\section{Methods (by class)}{
\itemize{
\item \code{plotSignificance(TMESplitResult)}: Method for \link{TMESplitResult}.

}}
\examples{
res <- new("TMESplitResult")
tryCatch(plotSignificance(res),
         error = function(e) message(conditionMessage(e)))
}
