% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/accessors.R
\name{programs}
\alias{programs}
\alias{programs,TMESplitResult-method}
\title{Accessor: per-program statistics}
\usage{
programs(x)

\S4method{programs}{TMESplitResult}(x)
}
\arguments{
\item{x}{A \link{TMESplitResult}.}
}
\value{
A \link[S4Vectors:DataFrame-class]{S4Vectors::DataFrame} with one row per program.
}
\description{
Accessor: per-program statistics
}
\section{Methods (by class)}{
\itemize{
\item \code{programs(TMESplitResult)}: Method for \link{TMESplitResult}.

}}
\examples{
res <- new("TMESplitResult")
programs(res)
}
