% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{show,TMESplitResult-method}
\alias{show,TMESplitResult-method}
\title{Show method for TMESplitResult}
\usage{
\S4method{show}{TMESplitResult}(object)
}
\arguments{
\item{object}{A \link{TMESplitResult}.}
}
\value{
Invisibly returns \code{NULL}; called for its side effect (printing).
}
\description{
Show method for TMESplitResult
}
\examples{
show(new("TMESplitResult"))
}
