% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/accessors.R
\name{wShared}
\alias{wShared}
\alias{wShared,TMESplitResult-method}
\title{Accessor: shared-program W matrix}
\usage{
wShared(x)

\S4method{wShared}{TMESplitResult}(x)
}
\arguments{
\item{x}{A \link{TMESplitResult}.}
}
\value{
A numeric matrix of shape \verb{cell_types x k_shared} (possibly empty).
}
\description{
Accessor: shared-program W matrix
}
\section{Methods (by class)}{
\itemize{
\item \code{wShared(TMESplitResult)}: Method for \link{TMESplitResult}.

}}
\examples{
res <- new("TMESplitResult")
wShared(res)
}
