% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/accessors.R
\name{wSpecific}
\alias{wSpecific}
\alias{wSpecific,TMESplitResult-method}
\title{Accessor: group-specific W matrices}
\usage{
wSpecific(x)

\S4method{wSpecific}{TMESplitResult}(x)
}
\arguments{
\item{x}{A \link{TMESplitResult}.}
}
\value{
A named \code{list} of numeric matrices, one per group.
}
\description{
Accessor: group-specific W matrices
}
\section{Methods (by class)}{
\itemize{
\item \code{wSpecific(TMESplitResult)}: Method for \link{TMESplitResult}.

}}
\examples{
res <- new("TMESplitResult")
wSpecific(res)
}
