% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotting.R
\name{plotActivities}
\alias{plotActivities}
\alias{plotActivities,TMESplitResult-method}
\title{Plot patient-level program activities by group}
\usage{
plotActivities(x, ...)

\S4method{plotActivities}{TMESplitResult}(x, ...)
}
\arguments{
\item{x}{A \link{TMESplitResult}.}

\item{...}{Extra aesthetics.}
}
\value{
A \link[ggplot2:ggplot]{ggplot2::ggplot} object.
}
\description{
Plot patient-level program activities by group
}
\section{Methods (by class)}{
\itemize{
\item \code{plotActivities(TMESplitResult)}: Method for \link{TMESplitResult}.

}}
\examples{
res <- new("TMESplitResult")
tryCatch(plotActivities(res), error = function(e) message(conditionMessage(e)))
}
