% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotting.R
\name{plotNetwork}
\alias{plotNetwork}
\alias{plotNetwork,TMESplitResult-method}
\title{Plot program co-occurrence / cell-type network}
\usage{
plotNetwork(x, ...)

\S4method{plotNetwork}{TMESplitResult}(x, ...)
}
\arguments{
\item{x}{A \link{TMESplitResult}.}

\item{...}{Extra aesthetics.}
}
\value{
A \link[ggplot2:ggplot]{ggplot2::ggplot} object.
}
\description{
Plot program co-occurrence / cell-type network
}
\section{Methods (by class)}{
\itemize{
\item \code{plotNetwork(TMESplitResult)}: Method for \link{TMESplitResult}.

}}
\examples{
res <- new("TMESplitResult")
tryCatch(plotNetwork(res), error = function(e) message(conditionMessage(e)))
}
